Snapshots dos dicionarios do Censo 2022

Dicionarios de variaveis embarcados no pacote para que o painel escolha variaveis sem depender do censobr em runtime (censo_variaveis()). Sao o resultado de data-raw/dicionarios.R sobre censobr::data_dictionary() (censobr v1.0.0).

Usage

dict_tracts_2022

dict_microdata_2022

Format

data.frame com quatro colunas:

variavel

codigo da variavel, aceito no argumento variaveis das funcoes de agregacao (ex.: "V0001", "demografia_V01007").

descricao

rotulo em pt-BR vindo do dicionario do IBGE.

dataset

subbase a que a variavel pertence (Basico, Pessoas, Domicilios, ...; no microdado, pessoas, domicilios ou familias).

tema

tema do IBGE (ex.: "Caracteristicas do Setor", "Demografia") ou NA.

dict_tracts_2022 tem 3.508 linhas (agregados por setor) e dict_microdata_2022 tem 247 linhas (microdado publico de 2022).

An object of class data.frame with 3508 rows and 4 columns.

An object of class data.frame with 247 rows and 4 columns.

Source

https://ftp.ibge.gov.br/Censos/ via censobr::data_dictionary(2022, dataset = c("tracts", "microdata")).

See also

Examples

if (FALSE) { # \dontrun{
d <- censo_variaveis(2022, "tracts")
table(d$dataset)
} # }